fix(auth): keep service below architecture ceiling
