+++ mobile/lib/models/auth_state.dart
@@ -1,0 +1,11 @@
+// ABOUTME: Defines the lifecycle states exposed by AuthService.
+// ABOUTME: Keeps authentication state independent from service implementation.
+
+/// Authentication state for the user.
+enum AuthState {
+  unauthenticated,
+  awaitingTosAcceptance,
+  authenticated,
+  checking,
+  authenticating,
+}
