+++ mobile/lib/services/auth_service.dart
@@ -15,8 +15,11 @@ import 'package:nostr_key_manager/nostr_key_manager.dart'
     show SecureKeyContainer, SecureKeyStorage, SecureKeyStorageException;
 import 'package:nostr_sdk/nostr_sdk.dart';
 import 'package:openvine/constants/app_constants.dart';
+import 'package:openvine/models/account_restore_failed_exception.dart';
 import 'package:openvine/models/auth_result.dart';
 import 'package:openvine/models/auth_rpc_capability.dart';
+import 'package:openvine/models/auth_state.dart';
+import 'package:openvine/models/auth_user_profile.dart';
 import 'package:openvine/models/authentication_source.dart';
 import 'package:openvine/models/known_account.dart';
 import 'package:openvine/services/account_deletion_proof_signer.dart';
@@ -41,7 +44,10 @@ import 'package:unified_logger/unified_logger.dart';
 
 // AuthResult moved to its own file (#4741) so auth/ collaborators can return it
 // without a facade import cycle; external consumers keep importing it here.
+export 'package:openvine/models/account_restore_failed_exception.dart';
 export 'package:openvine/models/auth_result.dart';
+export 'package:openvine/models/auth_state.dart';
+export 'package:openvine/models/auth_user_profile.dart';
 export 'package:openvine/models/authentication_source.dart';
 // Discovery callback contracts moved with the orchestrator (#4741); external
 // consumers (NostrService wiring) keep importing them from this facade.
@@ -60,24 +66,6 @@ const _kSessionRecoveryAnchorKey = 'session_recovery_anchor_npub';
 
 const _accountDeletionSessionExpiryMargin = Duration(minutes: 10);
 
-/// Authentication state for the user
-enum AuthState {
-  /// User is not authenticated (no keys stored)
-  unauthenticated,
-
-  /// User has keys but hasn't accepted Terms of Service yet
-  awaitingTosAcceptance,
-
-  /// User is authenticated (has valid keys and accepted TOS)
-  authenticated,
-
-  /// Authentication state is being checked
-  checking,
-
-  /// Authentication is in progress (generating/importing keys)
-  authenticating,
-}
-
 /// Whether the account-deletion flow may start.
 ///
 /// Deletion publishes irreversible events before it can attempt the reversible
@@ -92,72 +80,6 @@ enum AccountDeletionReadiness {
   /// destroy the user's content and leave their account alive. A fresh sign-in
   /// resolves it. Nothing has been published.
   requiresReauthentication,
-}
-
-/// Thrown by [AuthService.signInForAccount] when a returning-user sign-in
-/// does not restore the requested account — for example when an
-/// `importedKeys`/`automatic` account's identity keys are missing from secure
-/// storage, when a fallback authenticates a different primary account, or when
-/// [AuthService] lands in [AuthState.awaitingTosAcceptance] after an internal
-/// session-setup failure.
-///
-/// Previously these paths returned normally, leaving the caller (WelcomeBloc)
-/// believing the sign-in succeeded while the router kept the user pinned to
-/// `/welcome` — an invisible login loop. Throwing lets the caller route the
-/// user to the full login flow instead. See #5195.
-class AccountRestoreFailedException implements Exception {
-  const AccountRestoreFailedException(
-    this.pubkeyHex,
-    this.resolvedState, {
-    this.resolvedPubkeyHex,
-  });
-
-  /// The account (hex pubkey) whose restore was attempted.
-  final String pubkeyHex;
-
-  /// The [AuthState] the service resolved to.
-  final AuthState resolvedState;
-
-  /// The account (hex pubkey) that became active, if any.
-  final String? resolvedPubkeyHex;
-
-  @override
-  String toString() =>
-      'AccountRestoreFailedException: sign-in for $pubkeyHex resolved to '
-      '$resolvedState'
-      '${resolvedPubkeyHex == null ? '' : ' as $resolvedPubkeyHex'} '
-      'instead of the requested account';
-}
-
-/// User profile information
-class UserProfile {
-  const UserProfile({
-    required this.npub,
-    required this.publicKeyHex,
-    required this.displayName,
-    this.keyCreatedAt,
-    this.lastAccessAt,
-    this.about,
-    this.picture,
-    this.nip05,
-  });
-
-  /// Create minimal profile from secure key container
-  factory UserProfile.fromSecureContainer(SecureKeyContainer keyContainer) =>
-      UserProfile(
-        npub: keyContainer.npub,
-        publicKeyHex: keyContainer.publicKeyHex,
-        displayName: keyContainer.npub,
-      );
-
-  final String npub;
-  final String publicKeyHex;
-  final DateTime? keyCreatedAt;
-  final DateTime? lastAccessAt;
-  final String displayName;
-  final String? about;
-  final String? picture;
-  final String? nip05;
 }
 
 /// Callback to pre-fetch following list from REST API before auth state is set.
@@ -4540,10 +4462,8 @@ class AuthService implements BackgroundAwareService, BlockListSigner {
   /// this seam's parameter rather than the production call site's choice, so
   /// it cannot fail when that call site stops passing it.
   @visibleForTesting
-  void debugSetKeycastSigner(
-    KeycastRpc? signer, {
-    bool closePrevious = true,
-  }) => _setKeycastSigner(signer, closePrevious: closePrevious);
+  void debugSetKeycastSigner(KeycastRpc? signer, {bool closePrevious = true}) =>
+      _setKeycastSigner(signer, closePrevious: closePrevious);
 
   /// Test seam that lets unit tests install a [SecureKeyContainer] so
   /// `signOut`'s account-scoped invalidation (and any other code path
