+++ mobile/scripts/baseline/service_god_file_sizes.txt
@@ -6,7 +6,7 @@
 # This hard service baseline is authoritative over the advisory file_sizes.txt
 # baseline for service god-file ceilings.
 # Regenerate after shrinking/removing: UPDATE_BASELINE=1 bash scripts/check_service_god_file_ceiling.sh
-lib/services/auth_service.dart	4735
+lib/services/auth_service.dart	4674
 lib/services/curated_list_service.dart	1848
 lib/services/upload_manager.dart	2317
 lib/services/video_editor/video_editor_render_service.dart	1539
